Split CliContext to load config from memory, and consume Shipyard as a library
