+++ shipyard.py
@@ -39,6 +39,7 @@ except ImportError:
 # non-stdlib imports
 import click
 # local imports
+from convoy.context import Context
 import convoy.fleet
 import convoy.settings
 import convoy.util
@@ -49,28 +50,8 @@ logger = logging.getLogger('shipyard')
 _CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help'])
 
 
-class CliContext(object):
+class CliContext(Context):
     """CliContext class: holds context for CLI commands"""
-    def __init__(self):
-        """Ctor for CliContext"""
-        self.verbose = False
-        self.yes = False
-        self.config = None
-        self.batch_client = None
-        self.blob_client = None
-        self.queue_client = None
-        self.table_client = None
-        self.keyvault_client = None
-        # aad/keyvault options
-        self.keyvault_uri = None
-        self.keyvault_credentials_secret_id = None
-        self.aad_directory_id = None
-        self.aad_application_id = None
-        self.aad_auth_key = None
-        self.aad_user = None
-        self.aad_password = None
-        self.aad_cert_private_key = None
-        self.add_cert_thumbprint = None
 
     def initialize(self, creds_only=False, no_config=False):
         # type: (CliContext, bool, bool) -> None
@@ -79,23 +60,21 @@ class CliContext(object):
         :param bool creds_only: credentials only initialization
         :param bool no_config: do not configure context
         """
-        self._read_credentials_config()
-        self.keyvault_client = convoy.fleet.create_keyvault_client(
-            self, self.config)
-        del self.aad_directory_id
-        del self.aad_application_id
-        del self.aad_auth_key
-        del self.aad_user
-        del self.aad_password
-        del self.aad_cert_private_key
-        del self.aad_cert_thumbprint
-        self.config = None
-        self._init_config(creds_only)
-        if no_config:
-            return
+
+        obj_credentials = self._read_credentials_config()
+        obj_config, obj_pool, obj_jobs = None, None, None
         if not creds_only:
-            clients = convoy.fleet.initialize(self.config)
-            self._set_clients(*clients)
+            obj_config = self._read_config_config()
+            obj_pool = self._read_pool_config()
+            obj_jobs = self._read_obj_jobs()
+
+        super(Context, self).initialize(obj_credentials, obj_config, obj_pool, obj_jobs, creds_only, no_config)
+
+        # free mem
+        del self.json_credentials
+        del self.json_config
+        del self.json_pool
+        del self.json_jobs
 
     def _read_json_file(self, json_file):
         # type: (CliContext, pathlib.Path) -> None
@@ -106,11 +85,7 @@ class CliContext(object):
         """
         try:
             with json_file.open('r') as f:
-                if self.config is None:
-                    self.config = json.load(f)
-                else:
-                    self.config = convoy.util.merge_dict(
-                        self.config, json.load(f))
+                return json.load(f)
         except ValueError:
             raise ValueError(
                 ('Detected invalid JSON in file: {}. Please ensure the JSON '
@@ -130,103 +105,38 @@ class CliContext(object):
                 not isinstance(self.json_credentials, pathlib.Path)):
             self.json_credentials = pathlib.Path(self.json_credentials)
         if self.json_credentials.exists():
-            self._read_json_file(self.json_credentials)
+            return self._read_json_file(self.json_credentials)
 
-    def _init_config(self, creds_only=False):
-        # type: (CliContext, bool) -> None
-        """Initializes configuration of the context
-        :param CliContext self: this
-        :param bool creds_only: credentials only initialization
-        """
-        # use configdir if available
+    def _read_config_confg(self):
         if self.configdir is not None:
-            if self.json_credentials is None:
-                self.json_credentials = pathlib.Path(
-                    self.configdir, 'credentials.json')
             if self.json_config is None:
                 self.json_config = pathlib.Path(
                     self.configdir, 'config.json')
+        if self.json_config is None:
+            raise ValueError('config json was not specified')
+        elif not isinstance(self.json_config, pathlib.Path):
+            self.json_config = pathlib.Path(self.json_config)
+        return self._read_json_file(self.json_config)
+
+    def _read_pool_config(self):
+        if self.configdir is not None:
             if self.json_pool is None:
                 self.json_pool = pathlib.Path(self.configdir, 'pool.json')
+        if self.json_pool is None:
+            raise ValueError('pool json was not specified')
+        elif not isinstance(self.json_pool, pathlib.Path):
+            self.json_pool = pathlib.Path(self.json_pool)
+        return self._read_json_file(self.json_pool)
+
+    def _read_jobs_confg(self):
+        if self.configdir is not None:
             if self.json_jobs is None:
                 self.json_jobs = pathlib.Path(self.configdir, 'jobs.json')
-        # check for required json files
-        if (self.json_credentials is not None and
-                not isinstance(self.json_credentials, pathlib.Path)):
-            self.json_credentials = pathlib.Path(self.json_credentials)
-        if not creds_only:
-            if self.json_config is None:
-                raise ValueError('config json was not specified')
-            elif not isinstance(self.json_config, pathlib.Path):
-                self.json_config = pathlib.Path(self.json_config)
-            if self.json_pool is None:
-                raise ValueError('pool json was not specified')
-            elif not isinstance(self.json_pool, pathlib.Path):
-                self.json_pool = pathlib.Path(self.json_pool)
-        # fetch credentials from keyvault, if json file is missing
-        kvcreds = None
-        if self.json_credentials is None or not self.json_credentials.exists():
-            kvcreds = convoy.fleet.fetch_credentials_json_from_keyvault(
-                self.keyvault_client, self.keyvault_uri,
-                self.keyvault_credentials_secret_id)
-        # read credentials json, perform special keyvault processing if
-        # required sections are missing
-        if kvcreds is None:
-            self._read_json_file(self.json_credentials)
-            kv = convoy.settings.credentials_keyvault(self.config)
-            self.keyvault_uri = self.keyvault_uri or kv.keyvault_uri
-            self.keyvault_credentials_secret_id = (
-                self.keyvault_credentials_secret_id or
-                kv.keyvault_credentials_secret_id
-            )
-            if self.keyvault_credentials_secret_id is not None:
-                try:
-                    convoy.settings.credentials_batch(self.config)
-                    if len(list(convoy.settings.iterate_storage_credentials(
-                            self.config))) == 0:
-                        raise KeyError()
-                except KeyError:
-                    # fetch credentials from keyvault
-                    self.config = \
-                        convoy.fleet.fetch_credentials_json_from_keyvault(
-                            self.keyvault_client, self.keyvault_uri,
-                            self.keyvault_credentials_secret_id)
-        else:
-            self.config = kvcreds
-        del kvcreds
-        del self.keyvault_credentials_secret_id
-        # parse any keyvault secret ids from credentials
-        convoy.fleet.fetch_secrets_from_keyvault(
-            self.keyvault_client, self.config)
-        # read rest of config files
-        if not creds_only:
-            self._read_json_file(self.json_config)
-            self._read_json_file(self.json_pool)
-            if self.json_jobs is not None:
-                if not isinstance(self.json_jobs, pathlib.Path):
-                    self.json_jobs = pathlib.Path(self.json_jobs)
-                if self.json_jobs.exists():
-                    self._read_json_file(self.json_jobs)
-        # set internal config kv pairs
-        self.config['_verbose'] = self.verbose
-        self.config['_auto_confirm'] = self.yes
-        if self.verbose:
-            logger.debug('config:\n' + json.dumps(self.config, indent=4))
-        # free mem
-        del self.json_credentials
-        del self.json_config
-        del self.json_pool
-        del self.json_jobs
-        del self.verbose
-        del self.yes
-
-    def _set_clients(
-            self, batch_client, blob_client, queue_client, table_client):
-        """Sets clients for the context"""
-        self.batch_client = batch_client
-        self.blob_client = blob_client
-        self.queue_client = queue_client
-        self.table_client = table_client
+        if self.json_jobs is not None:
+            if not isinstance(self.json_jobs, pathlib.Path):
+                self.json_jobs = pathlib.Path(self.json_jobs)
+            if self.json_jobs.exists():
+                return self._read_json_file(self.json_jobs)
 
 
 # create a pass decorator for shared context between commands
